requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
networkx>=3.0
plotly>=5.17.0
python-dateutil>=2.8.2
pytest>=7.4.0
pytest-cov>=4.1.0

